Fix compile error with fmt 8.1.1
The error is:

    .../include/fmt/core.h:1728:7: error: static assertion failed:
    Cannot format an argument. To make type T formattable provide a
    formatter<T> specialization: https://fmt.dev/latest/api.html#udt

fmt 8.0.1 worked well. This is probably related to
https://github.com/fmtlib/fmt/issues/1841.

+++ src/session.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2021 Michal Sojka <michal.sojka@cvut.cz>
+// Copyright (C) 2021, 2022 Michal Sojka <michal.sojka@cvut.cz>
 // 
 // This file is part of boardproxy.
 // 
@@ -143,7 +143,7 @@ void Session::on_data_from_client(ev::io &w, int revents)
         break;
     }
     default:
-        logger->error("Unknown message type {}", h.id);
+        logger->error("Unknown message type {}", static_cast<int>(h.id));
     }
 
 //        switch (header.id) {
